Do not fail if config is not present in /etc
